@@ -1,56 +1,122 @@
 package org.example.controller;
 
 
-import com.example.hwswaggeruiind8.enitity.Faculty;
-import com.example.hwswaggeruiind8.enitity.Student;
-import com.example.hwswaggeruiind8.service.StudentService;
+
+
+import org.example.entity.Faculty;
+import org.example.entity.Student;
+import org.example.service.StudentService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-import java.util.Optional;
+
+import java.util.Collection;
 
 @RestController
-@RequestMapping("/student")
+@RequestMapping("student")
 public class StudentController {
-
     private final StudentService studentService;
+    Logger logger = LoggerFactory.getLogger(StudentController.class);
 
     public StudentController(StudentService studentService) {
         this.studentService = studentService;
     }
 
-    @PostMapping
-    public Student create(@RequestBody Student student) {
-        return studentService.add(student.getName(), student.getAge());
+    @GetMapping("{id}")
+    public ResponseEntity<Student> getStudent(@PathVariable Long id) {
+        Student student = studentService.findStudent(id);
+        if (student == null) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(student);
     }
 
-    @GetMapping
-    public Optional<Student> get(@RequestParam long id) {
-        return studentService.get(id);
+    @GetMapping("/age/{age}")
+    public ResponseEntity<Collection<Student>> getStudentByAge(@PathVariable int age) {
+        Collection<Student> result = studentService.getStudentsByAge(age);
+        if (result.size() == 0) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(result);
+    }
+
+    @PostMapping
+    public Student createStudent(@RequestBody Student student) {
+        return studentService.createStudent(student);
     }
 
     @PutMapping
-    public Student update(@RequestBody Student student) {
-        return studentService.update(student.getId(), student.getName(), student.getAge());
+    public ResponseEntity<Student> editStudent(@RequestBody Student student) {
+        Student foundStudent = studentService.editStudent(student);
+        if (foundStudent == null) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(foundStudent);
     }
 
-    @DeleteMapping
-    public Student delete(@RequestParam long id) {
-        return studentService.delete(id);
+    @DeleteMapping("{id}")
+    public ResponseEntity deleteStudent(@PathVariable Long id) {
+        studentService.deleteStudent(id);
+        return ResponseEntity.ok().build();
     }
 
-    @GetMapping("/by-age")
-    public List<Student> getByAge(@RequestParam int age) {
-        return studentService.getByAge(age);
+    @GetMapping("/age")
+    public ResponseEntity<Collection<Student>> getStudentByAgeBetween(@RequestParam int min, @RequestParam int max) {
+        Collection<Student> result = studentService.getStudentsByAgeBetween(min, max);
+        if (result.size() == 0) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(result);
     }
 
-    @GetMapping("/by-age-between")
-    public List<Student> getByAgeBetween(@RequestParam int min, @RequestParam int max) {
-        return studentService.getByAgeBetween(min, max);
+    @GetMapping("/faculty/{id}")
+    public ResponseEntity facultyByStudent(@PathVariable long id) {
+        Faculty foundFaculty = studentService.getStudentFaculty(id);
+        if (foundFaculty == null) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(foundFaculty);
     }
 
-    @GetMapping("/faculty-by-student-id")
-    public Faculty getFacultyByStudentId(@RequestParam Long id) {
-        return studentService.getFacultyByStudentId(id);
+    @GetMapping("/count")
+    public int getStudentCount() {
+        return studentService.getStudentCount();
+    }
+
+    @GetMapping("/avg-age")
+    public int getStudentAvgAge() {
+        return studentService.getStudentAvgAge();
+    }
+
+    @GetMapping("/last5")
+    public Collection<Student> getLast5Students() {
+        return studentService.getLast5Student();
+    }
+
+    @GetMapping("/name-list")
+    Collection<String> getNameList() {
+        return studentService.getNameList();
+    }
+
+    @GetMapping("/avg-age-by-stream")
+    public double getStudentAvgAgeByStream() {
+        return studentService.getStudentAvgAgeByStream();
+    }
+
+    @GetMapping("/get-number")
+    public int getNumber() {
+        return studentService.getNumber();
+    }
+
+    @GetMapping("/thread")
+    public void studentThread() {
+        studentService.studentThread();
+    }
+
+    @GetMapping("/synchronized-thread")
+    public void synchronizedStudentThread() throws InterruptedException {
+        studentService.synchronizedStudentThread();
     }
 }
